refactor(interpreter): keep remaining gas internal
